refactored utility files
